# backend/docs/openapi.yaml
openapi: 3.0.3
info:
  title: Kitchen Pantry CRM API
  description: |
    A comprehensive RESTful API for managing food service industry relationships, interactions, and sales opportunities. 
    Built on Node.js/Express with Supabase backend integration, supporting iPad-first CRM operations.
  version: 1.0.0
  contact:
    name: Kitchen Pantry CRM Support
    email: support@kitchenpantry.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: https://api.kitchenpantry.com/v1
    description: Production server
  - url: https://staging-api.kitchenpantry.com/v1
    description: Staging server
  - url: http://localhost:3000/api/v1
    description: Development server
security:
  - BearerAuth: []
paths:
...
  /opportunities:
    get:
      summary: List opportunities
      description: Retrieves opportunities with pipeline analysis and forecasting capabilities.
      tags:
        - Opportunities
      parameters:
        - name: page
          in: query
          description: Page number for pagination
          schema:
            type: integer
            default: 1
            minimum: 1
        - name: limit
          in: query
          description: Number of records per page
          schema:
            type: integer
            default: 50
            minimum: 1
            maximum: 100
        - name: organization_id
          in: query
          description: Filter opportunities by organization
          schema:
            type: string
            format: uuid
        - name: user_id
          in: query
          description: Filter opportunities by assigned user
          schema:
            type: string
            format: uuid
        - name: stage
          in: query
          description: Filter by sales stage
          schema:
            type: string
            enum: [prospecting, qualification, needs_analysis, proposal, negotiation, closed_won, closed_lost]
        - name: probability_min
          in: query
          description: Minimum probability threshold
          schema:
            type: integer
            minimum: 0
            maximum: 100
        - name: estimated_value_min
          in: query
          description: Minimum estimated value
          schema:
            type: number
            format: decimal
            minimum: 0
        - name: close_date_from
          in: query
          description: Filter by estimated close date range (from)
          schema:
            type: string
            format: date
        - name: close_date_to
          in: query
          description: Filter by estimated close date range (to)
          schema:
            type: string
            format: date
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OpportunityListResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
        '500':
          $ref: '#/components/responses/InternalServerError'

    post:
      summary: Create opportunity
      description: Creates a new sales opportunity with BANT qualification tracking.
      tags:
        - Opportunities
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OpportunityCreateRequest'
      responses:
        '201':
          description: Opportunity created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OpportunityResponse'
        '400':
          $ref: '#/components/responses/BadRequestError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
        '422':
          $ref: '#/components/responses/ValidationError'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /opportunities/{id}:
    get:
      summary: Get opportunity by ID
      description: Retrieves detailed information for a specific opportunity.
      tags:
        - Opportunities
      parameters:
        - name: id
          in: path
          required: true
          description: Opportunity ID
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OpportunityDetailResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
        '404':
          $ref: '#/components/responses/NotFoundError'
        '500':
          $ref: '#/components/responses/InternalServerError'

    put:
      summary: Update opportunity
      description: Updates an existing opportunity with complete resource replacement.
      tags:
        - Opportunities
      parameters:
        - name: id
          in: path
          required: true
          description: Opportunity ID
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OpportunityUpdateRequest'
      responses:
        '200':
          description: Opportunity updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OpportunityResponse'
        '400':
          $ref: '#/components/responses/BadRequestError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
        '404':
          $ref: '#/components/responses/NotFoundError'
        '422':
          $ref: '#/components/responses/ValidationError'
        '500':
          $ref: '#/components/responses/InternalServerError'

    patch:
      summary: Partially update opportunity
      description: Performs partial updates to an opportunity, updating only specified fields.
      tags:
        - Opportunities
      parameters:
        - name: id
          in: path
          required: true
          description: Opportunity ID
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OpportunityPatchRequest'
      responses:
        '200':
          description: Opportunity updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OpportunityResponse'
        '400':
          $ref: '#/components/responses/BadRequestError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
        '404':
          $ref: '#/components/responses/NotFoundError'
        '422':
          $ref: '#/components/responses/ValidationError'
        '500':
          $ref: '#/components/responses/InternalServerError'

    delete:
      summary: Delete opportunity
      description: Soft deletes an opportunity by setting is_deleted = true.
      tags:
        - Opportunities
      parameters:
        - name: id
          in: path
          required: true
          description: Opportunity ID
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: Opportunity deleted successfully
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
        '404':
          $ref: '#/components/responses/NotFoundError'
        '500':
          $ref: '#/components/responses/InternalServerError'
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token issued by Supabase Auth
  schemas:
...
    User:
      type: object
      properties:
        id:
          type: string
          format: uuid
        full_name:
          type: string
        email:
          type: string
          format: email
    Pagination:
      type: object
      properties:
        page:
          type: integer
          example: 1
        limit:
          type: integer
          example: 50
        total:
          type: integer
          example: 150
        pages:
          type: integer
          example: 3
        has_next:
          type: boolean
          example: true
        has_prev:
          type: boolean
          example: false
        next_cursor:
          type: string
          nullable: true
        prev_cursor:
          type: string
          nullable: true
    ResponseMetadata:
      type: object
      properties:
        timestamp:
          type: string
          format: date-time
        request_id:
          type: string
        pagination:
          $ref: '#/components/schemas/Pagination'

    # Organization schemas
...
    InteractionSummary:
      type: object
      properties:
        id:
          type: string
          format: uuid
        interaction_type:
          type: string
          enum: [CALL, EMAIL, MEETING, VISIT, TEXT, SOCIAL]
        interaction_date:
          type: string
          format: date-time
        subject:
          type: string
        outcome:
          type: string
          enum: [Positive, Neutral, Negative, Very Positive, Very Negative]
        user:
          type: object
          properties:
            full_name:
              type: string
...
    Opportunity:
      type: object
      properties:
        id:
          type: string
          format: uuid
        organization:
          type: object
          properties:
            id:
              type: string
              format: uuid
            name:
              type: string
            priority_level:
              type: string
              enum: [A, B, C, D]
        primary_contact:
          type: object
          properties:
            id:
              type: string
              format: uuid
            full_name:
              type: string
            job_title:
              type: string
        user:
          $ref: '#/components/schemas/User'
        opportunity_name:
          type: string
        description:
          type: string
        opportunity_type:
          type: string
          enum: [New Business, Existing Business, Renewal, Upsell, Cross-sell]
        stage:
          type: string
          enum: [prospecting, qualification, needs_analysis, proposal, negotiation, closed_won, closed_lost]
        probability:
          type: integer
          minimum: 0
          maximum: 100
        estimated_value:
          type: number
          format: decimal
        estimated_close_date:
          type: string
          format: date
        sales_cycle_days:
          type: integer
        lead_source:
          type: string
        budget_confirmed:
          type: boolean
        authority_confirmed:
          type: boolean
        need_confirmed:
          type: boolean
        timeline_confirmed:
          type: boolean
        proposal_submitted:
          type: boolean
        proposal_date:
          type: string
          format: date
        decision_criteria:
          type: string
        decision_timeline:
          type: string
        competitor_info:
          type: string
        priority_level:
          type: string
          enum: [A, B, C, D]
        recent_interactions:
          type: array
          items:
            $ref: '#/components/schemas/InteractionSummary'
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
...
    OpportunityCreateRequest:
      type: object
      properties:
        organization_id:
          type: string
          format: uuid
        primary_contact_id:
          type: string
          format: uuid
        opportunity_name:
          type: string
          minLength: 1
          maxLength: 255
        description:
          type: string
        opportunity_type:
          type: string
          enum: [New Business, Existing Business, Renewal, Upsell, Cross-sell]
        stage:
          type: string
          enum: [prospecting, qualification, needs_analysis, proposal, negotiation, closed_won, closed_lost]
        probability:
          type: integer
          minimum: 0
          maximum: 100
        estimated_value:
          type: number
          format: decimal
          minimum: 0
        estimated_close_date:
          type: string
          format: date
        lead_source:
          type: string
        budget_confirmed:
          type: boolean
          default: false
        authority_confirmed:
          type: boolean
          default: false
        need_confirmed:
          type: boolean
          default: false
        timeline_confirmed:
          type: boolean
          default: false
        decision_criteria:
          type: string
        decision_timeline:
          type: string
        competitor_info:
          type: string
        priority_level:
          type: string
          enum: [A, B, C, D]
      required:
        - organization_id
        - primary_contact_id
        - opportunity_name
        - stage
        - estimated_value
        - estimated_close_date
    OpportunityUpdateRequest:
      $ref: '#/components/schemas/OpportunityCreateRequest'
    OpportunityPatchRequest:
      type: object
      properties:
        opportunity_name:
          type: string
          minLength: 1
          maxLength: 255
        description:
          type: string
        opportunity_type:
          type: string
          enum: [New Business, Existing Business, Renewal, Upsell, Cross-sell]
        stage:
          type: string
          enum: [prospecting, qualification, needs_analysis, proposal, negotiation, closed_won, closed_lost]
        probability:
          type: integer
          minimum: 0
          maximum: 100
        estimated_value:
          type: number
          format: decimal
          minimum: 0
        estimated_close_date:
          type: string
          format: date
        lead_source:
          type: string
        budget_confirmed:
          type: boolean
        authority_confirmed:
          type: boolean
        need_confirmed:
          type: boolean
        timeline_confirmed:
          type: boolean
        decision_criteria:
          type: string
        decision_timeline:
          type: string
        competitor_info:
          type: string
        priority_level:
          type: string
          enum: [A, B, C, D]
    OpportunityResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        data:
          $ref: '#/components/schemas/Opportunity'
        meta:
          $ref: '#/components/schemas/ResponseMetadata'
    OpportunityDetailResponse:
      $ref: '#/components/schemas/OpportunityResponse'
    OpportunityListResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        data:
          type: array
          items:
            $ref: '#/components/schemas/Opportunity'
        meta:
          allOf:
            - $ref: '#/components/schemas/ResponseMetadata'
            - type: object
              properties:
                pipeline_summary:
                  type: object
                  properties:
                    total_opportunities:
                      type: integer
                    total_value:
                      type: number
                      format: decimal
                    weighted_value:
                      type: number
                      format: decimal
                    average_deal_size:
                      type: number
                      format: decimal
                    average_sales_cycle:
                      type: integer

    # Search schemas
...
    ErrorResponse:
      type: object
      properties:
        success:
          type: boolean
          example: false
        errors:
          type: array
          items:
            type: object
            properties:
              code:
                type: string
              message:
                type: string
              field:
                type: string
              details:
                type: string
        meta:
          type: object
          properties:
            timestamp:
              type: string
              format: date-time
            request_id:
              type: string
            retry_after:
              type: integer
  responses:
    UnauthorizedError:
      description: Authentication required
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            errors:
              - code: UNAUTHORIZED
                message: Authentication required
                details: Valid JWT token required in Authorization header
    ForbiddenError:
      description: Access denied
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            errors:
              - code: FORBIDDEN
                message: Access denied
                details: Insufficient permissions for this resource
    NotFoundError:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            errors:
              - code: NOT_FOUND
                message: Resource not found
                details: The requested resource does not exist
    BadRequestError:
      description: Invalid request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            errors:
              - code: BAD_REQUEST
                message: Invalid request format
                details: Request body must be valid JSON
...
    ValidationError:
      description: Validation error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            errors:
              - code: VALIDATION_ERROR
                message: Invalid input data
                field: email_primary
                details: Email format is invalid
    InternalServerError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            errors:
              - code: INTERNAL_ERROR
                message: Internal server error
                details: An unexpected error occurred
...